refactor: remove console logs from check-in and habit actions,
