+++ src/lib/state/check-in-actions.ts
@@ -8,7 +8,6 @@ const now = new Date().toISOString();
 
 export const createCheckIn = (habitId: string, frequency: number) => {
   const id = uuidv4();
-  console.log('Create CheckIn', habitId, frequency);
   checkIns$[id].set({
     id,
     habit_id: habitId,
@@ -47,33 +46,4 @@ export const deleteCheckIn = (id: string) => {
     frequency: checkIns$[id].get()?.frequency || null,
     habit_id: checkIns$[id].get()?.habit_id || null,
   });
-};
-
-type TCheckIn = Record<
-  string,
-  {
-    id: string;
-    checked_at: string | null;
-    created_at: string | null;
-    deleted: boolean | null;
-    frequency: number | null;
-    habit_id: string | null;
-    updated_at: string | null;
-  }
->;
-
-export const getTodaysCheckInId = (habitId: string, checkIns: TCheckIn) => {
-  const today = new Date();
-  if (!checkIns) return null;
-  const todaysCheckInId = Object.keys(checkIns).find((checkIn) => {
-    const checkInDate = checkIns[checkIn].checked_at;
-    if (!checkInDate) return false;
-    const date = new Date(checkInDate);
-    return (
-      date.toDateString() === today.toDateString() &&
-      checkIns[checkIn].habit_id === habitId
-    );
-  });
-
-  return todaysCheckInId;
 };
