+++ src/lib/state/habits-actions.ts
@@ -14,7 +14,6 @@ function formatTimeForSupabase(date: Date | string | null): string {
 
 export const createHabit = (habit: HabitWithoutId) => {
   const id = uuidv4();
-  console.log('Create Habit', habit);
 
   // Create the habit
   habits$[id].set({
